Update README.md
Added acronym glossary and team information for the project after discussion at Retro on 7 December. Need to check it's ok to put our team member names down here as it is a public repo? Or perhaps it would be better in a Wiki, although I don't see the option to create a wiki on my view of the repo.